upgrade contract-build and contract-sdk
