+++ packages/contracts-build/tsconfig.json
@@ -3,7 +3,6 @@
     "compilerOptions": {
         "baseUrl": ".",
         "outDir": "build",
-        "declaration": true,
         "rootDir": "src"
     },
     "include": [
