+++ packages/contracts-sdk/src/MarginedEngine.types.ts
@@ -1,4 +1,4 @@
-import {Uint128, Direction, Addr, Integer, AssetInfo, Boolean} from "./types";
+import {Uint128, Addr, Boolean} from "./types";
 export interface InstantiateMsg {
   eligible_collateral: string;
   fee_pool: string;
@@ -181,6 +181,7 @@ export type PositionFilter = "none" | {
 };
 export type PnlCalcOption = "spot_price" | "twap" | "oracle";
 export interface MigrateMsg {}
+export type Direction = "add_to_amm" | "remove_from_amm";
 export type ArrayOfPosition = Position[];
 export interface Position {
   block_time: number;
@@ -198,6 +199,19 @@ export interface Position {
   trader: Addr;
   vamm: Addr;
 }
+export interface Integer {
+  negative: boolean;
+  value: Uint128;
+}
+export type AssetInfo = {
+  token: {
+    contract_addr: Addr;
+  };
+} | {
+  native_token: {
+    denom: string;
+  };
+};
 export interface ConfigResponse {
   decimals: Uint128;
   eligible_collateral: AssetInfo;
@@ -217,7 +231,7 @@ export interface HooksResponse {
   hooks: string[];
 }
 export interface LastPositionIdResponse {
-  last_order_id: number;
+  last_position_id: number;
 }
 export interface StateResponse {
   bad_debt: Uint128;
