+++ packages/contracts-sdk/src/MarginedVamm.types.ts
@@ -1,4 +1,4 @@
-import {Uint128, Direction, Addr, Boolean, Integer} from "./types";
+import {Uint128, Addr, Boolean} from "./types";
 export interface InstantiateMsg {
   base_asset: string;
   base_asset_reserve: Uint128;
@@ -51,6 +51,7 @@ export type ExecuteMsg = {
     open: boolean;
   };
 };
+export type Direction = "add_to_amm" | "remove_from_amm";
 export type QueryMsg = {
   config: {};
 } | {
@@ -141,4 +142,8 @@ export interface StateResponse {
   open: boolean;
   quote_asset_reserve: Uint128;
   total_position_size: Integer;
+}
+export interface Integer {
+  negative: boolean;
+  value: Uint128;
 }
